solucion-VentaStock/requirements.txt
Django==2.1.7
django-extensions==2.1.6
django-crispy-forms==1.7.2
djangorestframework==3.9.1
django-rest-framework-client==0.1.1
django-cors-headers==2.5.2
dj-database-url==0.5.0
cx-Oracle==7.1.2
graphviz==0.10.1
social-auth-app-django==3.1.0
django-pwa==1.0.4
python-decouple==2.3
Unipath==1.1
Pillow==5.4.1
nose==1.3.7
pydotplus==2.0.2
South==1.0.2